# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "AInterviewer",
        "version": "0.0.11"
    },
    "paths": {
        "/api/projects/{project_id}/analysis/categories": {
            "get": {
                "tags": [
                    "analysis"
                ],
                "summary": "Get Analysis Categories",
                "operationId": "get_analysis_categories",
                "security": [
                    {
                        "APIKeyCookie": []
                    }
                ],
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid4",
                            "title": "Project Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AnalysisCategoryPublic"
                                    },
                                    "title": "Response Get Analysis Categories"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "analysis"
                ],
                "summary": "Create Analysis Category",
                "operationId": "create_analysis_category",
                "security": [
                    {
                        "APIKeyCookie": []
                    }
                ],
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid4",
                            "title": "Project Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnalysisCategoryCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/analysis/categories/{category_id}": {
            "put": {
                "tags": [
                    "analysis"
                ],
                "summary": "Update Analysis Category",
                "operationId": "update_analysis_category",
                "security": [
                    {
                        "APIKeyCookie": []
                    }
                ],
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid4",
                            "title": "Category Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnalysisCategoryCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnalysisCategoryPublic"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "analysis"
                ],
                "summary": "Delete Analysis Category",
                "operationId": "delete_analysis_category",
                "security": [
                    {
                        "APIKeyCookie": []
                    }
                ],
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid4",
                            "title": "Category Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/messages/{message_id}/annotations": {
            "get": {
                "tags": [
                    "analysis"
                ],
                "summary": "Get Message Annotations",
                "operationId": "get_message_annotations",
                "security": [
                    {
                        "APIKeyCookie": []
                    }
                ],
                "parameters": [
                    {
                        "name": "message_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid4",
                            "title": "Message Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MessageAnnotationPublic"
                                    },
                                    "title": "Response Get Message Annotations"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "analysis"
                ],
                "summary": "Add Message Annotation",
                "operationId": "add_message_annotation",
                "security": [
                    {
                        "APIKeyCookie": []
                    }
                ],
                "parameters": [
                    {
                        "name": "message_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid4",
                            "title": "Message Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MessageAnnotationCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageAnnotationPublic"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/analysis/annotations/{annotation_id}": {
            "put": {
                "tags": [
                    "analysis"
                ],
                "summary": "Update Message Annotation",
                "operationId": "update_message_annotation",
                "security": [
                    {
                        "APIKeyCookie": []
                    }
                ],
                "parameters": [
                    {
                        "name": "annotation_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid4",
                            "title": "Annotation Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MessageAnnotationCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageAnnotationPublic"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "analysis"
                ],
                "summary": "Delete Message Annotation",
                "operationId": "delete_message_annotation",
                "security": [
                    {
                        "APIKeyCookie": []
                    }
                ],
                "parameters": [
                    {
                        "name": "annotation_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid4",
                            "title": "Annotation Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AnalysisCategoryCreate": {
                "properties": {
                    "project_id": {
                        "type": "string",
                        "format": "uuid4",
                        "title": "Project Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "type": {
                        "$ref": "#/components/schemas/AnnotationType"
                    },
                    "color": {
                        "type": "string",
                        "title": "Color"
                    },
                    "min_value": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Min Value"
                    },
                    "max_value": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Max Value"
                    }
                },
                "type": "object",
                "required": [
                    "project_id",
                    "name",
                    "type",
                    "color"
                ],
                "title": "AnalysisCategoryCreate"
            },
            "AnalysisCategoryPublic": {
                "properties": {
                    "project_id": {
                        "type": "string",
                        "format": "uuid4",
                        "title": "Project Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "type": {
                        "$ref": "#/components/schemas/AnnotationType"
                    },
                    "color": {
                        "type": "string",
                        "title": "Color"
                    },
                    "min_value": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Min Value"
                    },
                    "max_value": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Max Value"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid4",
                        "title": "Id"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "project_id",
                    "name",
                    "type",
                    "color",
                    "id",
                    "created_at"
                ],
                "title": "AnalysisCategoryPublic"
            },
            "AnnotationType": {
                "type": "string",
                "enum": [
                    "tag",
                    "score"
                ],
                "title": "AnnotationType"
            },
            "AnnotationValueCreate": {
                "properties": {
                    "category_id": {
                        "type": "string",
                        "format": "uuid4",
                        "title": "Category Id"
                    },
                    "value_int": {
                        "type": "integer",
                        "title": "Value Int"
                    }
                },
                "type": "object",
                "required": [
                    "category_id",
                    "value_int"
                ],
                "title": "AnnotationValueCreate"
            },
            "AnnotationValuePublic": {
                "properties": {
                    "category_id": {
                        "type": "string",
                        "format": "uuid4",
                        "title": "Category Id"
                    },
                    "value_int": {
                        "type": "integer",
                        "title": "Value Int"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid4",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "category_id",
                    "value_int",
                    "id"
                ],
                "title": "AnnotationValuePublic"
            },
...
            "ErrorResponse": {
                "properties": {
                    "detail": {
                        "type": "string",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "required": [
                    "detail"
                ],
                "title": "ErrorResponse"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "MessageAnnotationCreate": {
                "properties": {
                    "message_id": {
                        "type": "string",
                        "format": "uuid4",
                        "title": "Message Id"
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid4",
                        "title": "User Id"
                    },
                    "comment": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Comment"
                    },
                    "values": {
                        "items": {
                            "$ref": "#/components/schemas/AnnotationValueCreate"
                        },
                        "type": "array",
                        "title": "Values"
                    }
                },
                "type": "object",
                "required": [
                    "message_id",
                    "user_id",
                    "values"
                ],
                "title": "MessageAnnotationCreate"
            },
            "MessageAnnotationPublic": {
                "properties": {
                    "message_id": {
                        "type": "string",
                        "format": "uuid4",
                        "title": "Message Id"
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid4",
                        "title": "User Id"
                    },
                    "comment": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Comment"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid4",
                        "title": "Id"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    },
                    "values": {
                        "items": {
                            "$ref": "#/components/schemas/AnnotationValuePublic"
                        },
                        "type": "array",
                        "title": "Values"
                    }
                },
                "type": "object",
                "required": [
                    "message_id",
                    "user_id",
                    "id",
                    "created_at",
                    "updated_at",
                    "values"
                ],
                "title": "MessageAnnotationPublic"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "APIKeyCookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "token"
            }
        }
    }
}